requirements.txt
APScheduler==3.11.1
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.1
Flask==3.1.2
Flask-APScheduler==1.13.1
flask-cors==6.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
PyMySQL==1.1.2
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
requests==2.32.5
six==1.17.0
tomli==2.3.0
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
